pyproject.toml
[project]
name = "my-drive-manager"
version = "0.1.0"
description = "Controlling my backup files"
authors = [
    { name = "Raphael Silva do Nascimento", email = "eng.raphaelsn@gmail.com" },
]
keywords = ["cli", "drive", "manager"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aio-pika>=9.5.5",
    "ffmpeg-python>=0.2.0",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "moviepy>=2.2.1",
    "pillow>=11.2.1",
    "pydantic>=2.11.6",
    "pytest>=8.3.5",
    "pymediainfo>=7.0.1",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.25.3",
    "pytest>=8.3.5",
]
